fix(edgeless): horizontal pan with mouse under the windows system
